Add more code

+++ api/v2/model.go
@@ -213,8 +213,8 @@ type ReplicaConfig struct {
 	Filter                       *FilterConfig              `json:"filter,omitempty"`
 	Mounter                      *MounterConfig             `json:"mounter,omitempty"`
 	Sink                         *SinkConfig                `json:"sink,omitempty"`
-	EventCollectorBatchCount     *int                       `json:"event_collector_batch_count,omitempty"`
-	EventCollectorBatchBytes     *int                       `json:"event_collector_batch_bytes,omitempty"`
+	EventCollectorBatchCount     *uint64                    `json:"event_collector_batch_count,omitempty"`
+	EventCollectorBatchBytes     *uint64                    `json:"event_collector_batch_bytes,omitempty"`
 	Consistent                   *ConsistentConfig          `json:"consistent,omitempty"`
 	Scheduler                    *ChangefeedSchedulerConfig `json:"scheduler,omitempty"`
 	Integrity                    *IntegrityConfig           `json:"integrity,omitempty"`
@@ -643,14 +643,14 @@ func ToAPIReplicaConfig(c *config.ReplicaConfig) *ReplicaConfig {
 	cloned := c.Clone()
 
 	res := &ReplicaConfig{
-		MemoryQuota:           cloned.MemoryQuota,
-		CaseSensitive:         cloned.CaseSensitive,
-		ForceReplicate:        cloned.ForceReplicate,
-		IgnoreIneligibleTable: cloned.IgnoreIneligibleTable,
-		CheckGCSafePoint:      cloned.CheckGCSafePoint,
-		EnableSyncPoint:       cloned.EnableSyncPoint,
-		EnableTableMonitor:    cloned.EnableTableMonitor,
-		BDRMode:               cloned.BDRMode,
+		MemoryQuota:              cloned.MemoryQuota,
+		CaseSensitive:            cloned.CaseSensitive,
+		ForceReplicate:           cloned.ForceReplicate,
+		IgnoreIneligibleTable:    cloned.IgnoreIneligibleTable,
+		CheckGCSafePoint:         cloned.CheckGCSafePoint,
+		EnableSyncPoint:          cloned.EnableSyncPoint,
+		EnableTableMonitor:       cloned.EnableTableMonitor,
+		BDRMode:                  cloned.BDRMode,
 		EventCollectorBatchCount: cloned.EventCollectorBatchCount,
 		EventCollectorBatchBytes: cloned.EventCollectorBatchBytes,
 	}
